requirements.txt
langchain_huggingface
langchain
langchain-community
langchain-core
langgraph
trafilatura
python-dotenv